Sort the radar targets at the beginning of the game
The players' order in the radar targets list doesn't usually match the network spawning order which radar targets sort by when a radar booster is activated. This causes the target order to change when a radar booster is activated, and after doing so the target may not switch in the expected way.

Sorting the players at the start of the game based on this network spawning keeps the player targets' order consistent before and after a radar booster is activated.