Set the correct `TransformAndName` names to players' usernames on join
Since the list of `TransformAndName` instances is sorted not by the index in the `StartOfRound.allPlayerScripts` array, but the network ID, the vanilla code would set the name of the wrong transform in the list after sorting. Correct that by matching a player instance to its transform in the list.