Use ChangeNameOfTargetTransform when setting player names instead
Using a hand-rolled function for finding the player's transform index is totally unnecessary when the base game already has a function to do what we want. This way, we can react to the name being set using the same injection that was used to update radar boosters' names on the map.