Use a patch to enable the terminal map based on visibility
This eschews the TerminalVisibilityTracker that would enable the camera based simply on whether the terminal was open. Instead, now that Darmuh's terminal mod takes control of the enabled state of the map, we don't need to worry about the overall state of the terminal, but only the vanilla map image component.